test: fold slotted island fixture into server-components
