--- test/fixtures/server-component-slotted/nuxt.config.ts
@@ -1,8 +1,0 @@
-import { isNuxtPrepare, projectSuffix, withMatrix } from '../../matrix.ts'
-
-export default withMatrix({
-  ...(isNuxtPrepare ? {} : { buildDir: `.nuxt-${projectSuffix}` }),
-  experimental: {
-    componentIslands: true,
-  },
-})
