--- test/fixtures/server-component-slotted/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "fixture-server-component-slotted",
-  "private": true,
-  "type": "module",
-  "dependencies": {
-    "nuxt": "workspace:*"
-  },
-  "engines": {
-    "node": "^22.19.0 || ^24.11.0 || >=26.0.0"
-  }
-}
