+++ test/fixtures/server-components/app/components/SlottedStyles.server.vue
