+++ test/fixtures/server-components/app/pages/slotted-styles.vue
